reformat(web): move externalStore to @safe-global/utils
